+++ R/draw.R
@@ -15,9 +15,8 @@
 #'          Firefox is an exception. See Details for instructions on how to enable this
 #'          behaviour in Firefox.
 #' @param title \code{string} to customize the title of the UI window.
-#' @param editor \code{character} either "leaflet.extras" or "leafpm"
+#' @param editor \code{character} currently, only "leafpm" is supported.
 #' @param editorOptions \code{list} of options suitable for passing to
-#'     either \code{leaflet.extras::addDrawToolbar} or
 #'     \code{leafpm::addPmToolbar}.
 #' @param ... additional arguments passed on to \code{\link{editMap}}.
 #'
@@ -37,7 +36,7 @@ drawFeatures = function(map = NULL,
                         record = FALSE,
                         viewer = shiny::paneViewer(),
                         title = "Draw Features",
-                        editor = c("leaflet.extras", "leafpm"),
+                        editor = "leafpm",
                         editorOptions = list(),
                         ...) {
   res = editMap(x = map,