+++ R/editAttributes.R
@@ -224,7 +224,8 @@ editAttributes <- function(dat, zoomto = NULL, col_add = TRUE, reset = TRUE, pro
         id = "map",
         targetLayerId = 'editLayer',
         sf = TRUE,
-        editorOptions = list(editOptions = leaflet.extras::editToolbarOptions(edit = df$edit_logic)),
+        # editorOptions = list(editOptions = leaflet.extras::editToolbarOptions(edit = df$edit_logic)),
+        editorOptions = list(editOptions = leafpm::pmToolbarOptions(editMode = df$edit_logic))
       )
     })